agrupar aggregations en el mismo archivo
