+++ controllers/filters.controller.js
@@ -1,5 +1,6 @@
 const User = require('../models/users.model');
 const Topics = require('../models/topics.model');
+const Inscription = require('../models/inscriptions.model');
 
 exports.filterTopics = async (req, res) => {
   const { postalCode, topics } = req.body;
@@ -47,3 +48,38 @@ exports.filterTopics = async (req, res) => {
   }
 };
 
+//filter by zone added into the same file
+const getZonesWithMostInscriptions = (req, res, next) => {
+  (async () => {
+    const result = await Inscription.aggregate([
+      {
+        $lookup: {
+          from: 'users',
+          localField: 'user',
+          foreignField: '_id',
+          as: 'user',
+        },
+      },
+      {
+        $group: {
+          _id: '$user.postalCode',
+          totalInscriptions: { $sum: 1 },
+        },
+      },
+      {
+        $sort: { totalInscriptions: -1 },
+      },
+      {
+        $project: {
+          _id: 0,
+          postalCode: '$_id',
+          totalInscriptions: 1,
+        },
+      },
+    ]);
+  
+      res.status(200).json(result);
+    })().catch(next) 
+      console.log(error);
+      res.status(500).json({ error: 'Error de servidor ' });
+  }
